+++ cranko-rc-info-v1
[[projects]]
qnames = ["tectonic_xdv", "cargo"]
bump_spec = "micro bump"

[[projects]]
qnames = ["tectonic_errors", "cargo"]
bump_spec = "micro bump"

[[projects]]
qnames = ["tectonic_xetex_format", "cargo"]
bump_spec = "micro bump"

[[projects]]
qnames = ["tectonic_status_base", "cargo"]
bump_spec = "micro bump"

[[projects]]
qnames = ["tectonic_io_base", "cargo"]
bump_spec = "micro bump"

[[projects]]
qnames = ["tectonic_geturl", "cargo"]
bump_spec = "micro bump"

[[projects]]
qnames = ["tectonic_docmodel", "cargo"]
bump_spec = "micro bump"

[[projects]]
qnames = ["tectonic_dep_support", "cargo"]
bump_spec = "micro bump"

[[projects]]
qnames = ["tectonic_bundles", "cargo"]
bump_spec = "micro bump"

[[projects]]
qnames = ["tectonic_bridge_graphite2", "cargo"]
bump_spec = "micro bump"

[[projects]]
qnames = ["tectonic_bridge_flate", "cargo"]
bump_spec = "micro bump"

[[projects]]
qnames = ["tectonic_bridge_core", "cargo"]
bump_spec = "micro bump"

[[projects]]
qnames = ["tectonic_xetex_layout", "cargo"]
bump_spec = "micro bump"

[[projects]]
qnames = ["tectonic_pdf_io", "cargo"]
bump_spec = "micro bump"

[[projects]]
qnames = ["tectonic_engine_xetex", "cargo"]
bump_spec = "micro bump"

[[projects]]
qnames = ["tectonic_engine_xdvipdfmx", "cargo"]
bump_spec = "micro bump"

[[projects]]
qnames = ["tectonic_engine_spx2html", "cargo"]
bump_spec = "minor bump"

[[projects]]
qnames = ["tectonic_engine_bibtex", "cargo"]
bump_spec = "micro bump"

+++

+++ crates/bridge_flate/src/lib.rs
@@ -72,7 +72,7 @@ pub unsafe extern "C" fn tectonic_flate_compress(
         (c.total_out(), FlateResult::Success)
     };
 
-    *output_len = size as u64;
+    *output_len = size;
     result
 }
 
@@ -101,7 +101,7 @@ pub unsafe extern "C" fn tectonic_flate_decompress(
         Err(_) => (0, FlateResult::OtherError),
     };
 
-    *output_len = size as u64;
+    *output_len = size;
     result
 }
